Populate contributors section with data from *all* repos (#59)
Previously, the contributors section was populated with data of
contributors from solely the jboss-outreach/gci repo. Modify the
code to load data from all repos under the jboss-outreach org.

Future improvements:
 * Add a loading indicator when the data is being populated
 * Further reduce loading time

**Reducing loading time:**
One way to reduce this is to pre-process and pre-cache the data with
a backend server that updates itself every few minutes or so. The client
then simply queries the server. This reduces the number of queries
from dozens to just one.

If there are any other possibilities, do inform me and I could look
into them. I'm quite new to this and would love to know more :)